Crate buck2_error

Modules

any

Integrations of buck2_error::Error with anyhow::Error and StdError.

classify

conversion

Conversion impls for different error types to ‘buck2_error::Error’

macros

source_location

starlark_error

Starlark::Error <-> buck2_error::Error conversion implementation.

Macros

buck2_error

internal_error

Indicates a bug in buck2.

Structs

Error

The core error type provided by this crate.

UniqueRootId

Uniquely identifies an instance of an error root

Enums

ErrorTag

See the documentation in the error.proto file for details. Error types are - by design - restricted to being set exactly once at the error definition site. While they are useful, that means that they are enough on their own to represent all the error metadata we want. Until we figure out what exactly a more complete model looks like, error tags a mechanism that offers maximum flexibility - they can be added to any error anywhere.

ExitCode

Common exit codes for buck with stronger semantic meanings

Tier

A piece of metadata to indicate whether this error is an infra or user error.

Traits

BuckErrorContext

Provides the context method for Result.

TypedContext

Functions

Ok

Allows simpler construction of the Ok case when the result type can’t be inferred.

provide_metadata

Provide metadata about an error.

Type Aliases

DynLateFormat

Result

Derive Macros

Error

Generates an error impl for the type.