buck2_error::any

Struct ProvidableMetadata

pub struct ProvidableMetadata {
    pub tags: Vec<ErrorTag>,
    pub string_tags: Vec<String>,
    pub source_location: SourceLocation,
    pub action_error: Option<ActionError>,
}

This can be provided by an error to inject buck2-specific information about it.

For action_error, and the source information, only the value that appears last in the source chain will be used. The derive macro typically handles this to prevent any surprises, however if this value is being provided manually then care may need to be taken.

Fields
tags: Vec<ErrorTag>string_tags: Vec<String>source_location: SourceLocationaction_error: Option<ActionError>

The protobuf ActionError, if the root was an action error