buck2_error

Module classify

Enums

ErrorSourceArea

Tier

Constants

ERROR_TAG_UNCLASSIFIED

When there’s no tag, but we want to put something in Scuba, we use this.

Traits

ErrorLike

ErrorTagExtra

Functions

best_error

Pick the most interesting error by best tag.

best_tag

Pick the most interesting tag from a list of tags.

source_area

tag_is_generic

Errors can be categorized by tags only if they have any non-generic tags.

tag_is_hidden

Hidden tags only used internally, for categorization.