buck2_error

Derive Macro Error

#[derive(Error)]
{
    // Attributes available to this derive:
    #[error]
    #[source]
    #[buck2]
}

Generates an error impl for the type.

This macro is a drop-in replacement for [thiserror::Error]. In the near future, all uses of thiserror in buck2/app will be replaced with this macro.

Currently, the only distinction from thiserror::Error is that an additional impl of AnyError is generated for the type, which makes some of the interactions with buck2_error more ergonomic. In the future, this macro will also be used to be able to annotate errors with additional structured context information.

Example

#[derive(Debug, buck2_error::Error)]
#[error("My error type")]
#[buck2(tag = buck2_error::ErrorTag::Input)]
struct MyError;

let e = buck2_error::Error::from(MyError);
assert_eq!(&format!("{}", e), "My error type");