buck2_error

Struct Error

pub struct Error(/* private fields */);

The core error type provided by this crate.

While this type has many of the features of anyhow::Error, in most places you should continue to use anyhow. This type is only expected to appear on a small number of APIs which require a clonable error.

Unlike anyhow::Error, this type supports no downcasting. That is an intentional choice - downcasting errors is fragile and becomes difficult to support in conjunction with anyhow compatibility.

Implementations

impl Error

pub fn new( error_msg: String, error_tag: ErrorTag, source_location: SourceLocation, action_error: Option<ActionError>, ) -> Self

pub fn action_error(&self) -> Option<&ActionError>

pub fn mark_emitted(self, late_format: Arc<DynLateFormat>) -> Self

pub fn is_emitted(&self) -> Option<impl Debug + Display + '_>

If the error has not been emitted yet, returns None, otherwise Some.

Most errors are only shown to the user once. However, some errors, specifically action errors, are shown to the user twice: Once when the error occurs, and again at the end of the build in the form of a short “Failed to build target” summary.

After the error has been shown to the user for the first time, it is marked as emitted. The late formatter that is returned here is what should be printed at the end of the build

pub fn get_stack_for_debug(&self) -> String

Only intended to be used for debugging, helps to understand the structure of the error

pub fn root_id(&self) -> UniqueRootId

Identifier for deduplication during a build.

pub fn category_key(&self) -> String

Stable identifier for grouping errors.

This tries to include the least information possible that can be used to uniquely identify an error type.

pub fn source_location(&self) -> &SourceLocation

pub fn context<C: Into<ContextValue>>(self, context: C) -> Self

pub fn string_tag(self, context: &str) -> Self

pub fn context_for_starlark_backtrace(self, context: StarlarkContext) -> Self

pub fn tag(self, tags: impl IntoIterator<Item = ErrorTag>) -> Self

pub fn get_tier(&self) -> Option<Tier>

pub fn exit_code(&self) -> ExitCode

pub fn find_typed_context<T: TypedContext>(&self) -> Option<Arc<T>>

pub fn string_tags(&self) -> Vec<String>

pub fn tags(&self) -> Vec<ErrorTag>

Get all the tags that have been added to this error

pub fn best_tag(&self) -> Option<ErrorTag>

The most interesting tag among this error tags.

pub fn has_tag(&self, tag: ErrorTag) -> bool