buck2_error

Trait BuckErrorContext

pub trait BuckErrorContext<T>: Sealed {
    // Required methods
    fn buck_error_context<C: Into<ContextValue>>(self, context: C) -> Result<T>;
    fn with_buck_error_context<C, F>(self, f: F) -> Result<T>
       where C: Into<ContextValue>,
             F: FnOnce() -> C;
    fn buck_error_context_anyhow<C: Into<ContextValue>>(
        self,
        context: C,
    ) -> Result<T>;
    fn with_buck_error_context_anyhow<C, F>(self, f: F) -> Result<T>
       where C: Into<ContextValue>,
             F: FnOnce() -> C;
    fn compute_context<TC: TypedContext, C1: Into<ContextValue>, C2: Into<ContextValue>, F: FnOnce(Arc<TC>) -> C1, F2: FnOnce() -> C2>(
        self,
        map_context: F,
        new_context: F2,
    ) -> Result<T>;

    // Provided methods
    fn tag(self, tag: ErrorTag) -> Result<T> { ... }
    fn internal_error(self, message: &str) -> Result<T> { ... }
    fn with_internal_error<F>(self, f: F) -> Result<T>
       where F: FnOnce() -> String { ... }
}

Provides the context method for Result.

This trait is analogous to the anyhow::Context trait. It is mostly a drop-in replacement, and in the near future, uses of anyhow::Context in buck2/app will be broadly replaced with use of this trait. Subsequently, additional APIs will be provided for annotating errors with structured context data.

Required Methods

fn buck_error_context<C: Into<ContextValue>>(self, context: C) -> Result<T>

fn with_buck_error_context<C, F>(self, f: F) -> Result<T>
where C: Into<ContextValue>, F: FnOnce() -> C,

fn buck_error_context_anyhow<C: Into<ContextValue>>( self, context: C, ) -> Result<T>

Code below returns an anyhow::Error, it is used while we transition from anyhow to buck2_error in buck2/app TODO(minglunli): Delete the code below once we have fully transitioned to buck2_error

fn with_buck_error_context_anyhow<C, F>(self, f: F) -> Result<T>
where C: Into<ContextValue>, F: FnOnce() -> C,

fn compute_context<TC: TypedContext, C1: Into<ContextValue>, C2: Into<ContextValue>, F: FnOnce(Arc<TC>) -> C1, F2: FnOnce() -> C2>( self, map_context: F, new_context: F2, ) -> Result<T>

Supports adding context to an error by either augmenting the most recent context if its the requested type or by adding a new context.

Provided Methods

fn tag(self, tag: ErrorTag) -> Result<T>

fn internal_error(self, message: &str) -> Result<T>

fn with_internal_error<F>(self, f: F) -> Result<T>
where F: FnOnce() -> String,

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.